Refactor: create src folder
in flutter, it's common to store classes inside a folder called src
