+++ lib/main.dart
@@ -2,16 +2,16 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
-import 'package:proyect_atenea/presentation/pages/auth/login_page.dart';
-import 'package:proyect_atenea/presentation/pages/auth/register_page.dart';
-import 'package:proyect_atenea/presentation/pages/auth/splash_page.dart';
-import 'package:proyect_atenea/presentation/pages/home/home_page.dart';
-import 'package:proyect_atenea/presentation/providers/session_provider.dart';
-import 'package:proyect_atenea/presentation/values/app_theme.dart';
-import 'package:proyect_atenea/data/data_sources/session_manager_data_source.dart';
-import 'package:proyect_atenea/data/repositories/session_repository_impl.dart';
-import 'package:proyect_atenea/domain/use_cases/session_use_cases.dart';
-import 'package:proyect_atenea/firebase_options.dart';
+import 'package:proyect_atenea/src/presentation/pages/auth/login_page.dart';
+import 'package:proyect_atenea/src/presentation/pages/auth/register_page.dart';
+import 'package:proyect_atenea/src/presentation/pages/auth/splash_page.dart';
+import 'package:proyect_atenea/src/presentation/pages/home/home_page.dart';
+import 'package:proyect_atenea/src/presentation/providers/session_provider.dart';
+import 'package:proyect_atenea/src/presentation/values/app_theme.dart';
+import 'package:proyect_atenea/src/data/data_sources/session_manager_data_source.dart';
+import 'package:proyect_atenea/src/data/repositories/session_repository_impl.dart';
+import 'package:proyect_atenea/src/domain/use_cases/session_use_cases.dart';
+import 'package:proyect_atenea/src/firebase_options.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
