+++ lib/src/domain/entities/session.dart
