+++ lib/src/presentation/pages/auth/login_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:proyect_atenea/presentation/values/app_theme.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_button.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_field.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_scaffold.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_switch.dart';
+import 'package:proyect_atenea/src/presentation/values/app_theme.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_button.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_field.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_scaffold.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_switch.dart';
 
 class LoginPage extends StatelessWidget {
   const LoginPage({
