+++ lib/src/presentation/pages/auth/register_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:proyect_atenea/presentation/values/app_theme.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_button.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_field.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_scaffold.dart';
+import 'package:proyect_atenea/src/presentation/values/app_theme.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_button.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_field.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_scaffold.dart';
 
 class RegisterPage extends StatelessWidget {
   const RegisterPage({
