+++ lib/src/presentation/widgets/atenea_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:proyect_atenea/presentation/values/app_theme.dart';
+import 'package:proyect_atenea/src/presentation/values/app_theme.dart';
 
 class AteneaButton extends StatelessWidget {
   final String? text;
