+++ lib/src/presentation/widgets/atenea_scaffold.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:proyect_atenea/presentation/widgets/atenea_background.dart';
+import 'package:proyect_atenea/src/presentation/widgets/atenea_background.dart';
 
 class AteneaScaffold extends StatelessWidget {
   final Widget body;
